Remove miral::WorkspacePolicy
