--- src/miral/workspace_policy.cpp
@@ -1,29 +1,0 @@
-/*
- * Copyright © 2017 Canonical Ltd.
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 or 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Authored by: Alan Griffiths <alan@octopull.co.uk>
- */
-
-#include <miral/workspace_policy.h>
-
-void miral::WorkspacePolicy::advise_adding_to_workspace(
-    std::shared_ptr<Workspace> const& , std::vector<Window> const&)
-{
-}
-
-void miral::WorkspacePolicy::advise_removing_from_workspace(
-    std::shared_ptr<Workspace> const&, std::vector<Window> const&)
-{
-}
